update dialog can add tracks to playlist

+++ MusicPlayer/app/src/main/java/denis/musicplayer/data/media/model/Album.kt
@@ -3,7 +3,7 @@ package denis.musicplayer.data.media.model
 /**
  * Created by denis on 01/01/2018.
  */
-data class Album(val id: Int,
+data class Album(val id: Long,
                  val album: String,
                  val artist: String,
                  val art: String?) {
