+++ MusicPlayer/app/src/main/java/denis/musicplayer/data/media/model/Artist.kt
@@ -3,7 +3,7 @@ package denis.musicplayer.data.media.model
 /**
  * Created by denis on 01/01/2018.
  */
-data class Artist(val id: Int,
+data class Artist(val id: Long,
                   val artist: String,
                   val albumCount: String,
                   val trackCount: String) {
