+++ MusicPlayer/app/src/main/java/denis/musicplayer/data/media/model/Genre.kt
@@ -3,7 +3,7 @@ package denis.musicplayer.data.media.model
 /**
  * Created by denis on 01/01/2018.
  */
-data class Genre(val id: Int,
+data class Genre(val id: Long,
                  val name: String) {
     override fun equals(other: Any?): Boolean {
         return id == (other as Genre).id
