+++ MusicPlayer/app/src/main/java/denis/musicplayer/data/playlist/model/Playlist.kt
@@ -3,5 +3,5 @@ package denis.musicplayer.data.playlist.model
 /**
  * Created by denis on 02/01/2018.
  */
-data class Playlist (val id: Int,
+data class Playlist (val id: Long,
                      val name: String)
